requirements.txt
streamlit>=1.32.0
streamlit-authenticator>=0.3.2
sxtwl>=2.0.3
google-genai>=1.5.0
weasyprint>=60.0
python-dotenv>=1.0.0
Pillow>=10.0.0
json-repair>=0.28.0
pyyaml>=6.0
